requirements.txt
Django==2.0.6
django-mdeditor==0.1.13
PyMySQL==0.9.3
pytz==2018.9
uWSGI==2.0.17.1
